requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24
pydantic>=2.6
python-dotenv>=1.0
jinja2>=3.1
requests>=2.31
python-multipart>=0.0.9
pypdf>=4.0
python-docx>=1.1
python-pptx>=0.6.23
